refactor: use error codes for recently opened error handling instead of comparing error messages
