+++ packages/renderer-worker/src/parts/IsString/IsString.js
@@ -1,0 +1,3 @@
+export const isString = (value) => {
+  return typeof value === 'string'
+}
