+++ packages/shared-process/src/parts/ErrorCodes/ErrorCodes.js
@@ -4,6 +4,7 @@ export const E_ICON_THEME_NOT_FOUND = 'E_ICON_THEME_NOT_FOUND'
 export const E_INCOMPATIBLE_NATIVE_MODULE = 'E_INCOMPATIBLE_NATIVE_MODULE'
 export const E_MANIFEST_NOT_FOUND = 'E_MANIFEST_NOT_FOUND'
 export const E_MODULES_NOT_SUPPORTED_IN_ELECTRON = 'E_MODULES_NOT_SUPPORTED_IN_ELECTRON'
+export const E_JSON_PARSE = 'E_JSON_PARSE'
 export const EACCES = 'EACCES'
 export const ECONNRESET = 'ECONNRESET'
 export const EEXIST = 'EEXIST'
