+++ packages/shared-process/src/parts/JsonParsingError/JsonParsingError.js
@@ -1,10 +1,12 @@
 import * as JoinLines from '../JoinLines/JoinLines.js'
 import * as SplitLines from '../SplitLines/SplitLines.js'
+import * as ErrorCodes from '../ErrorCodes/ErrorCodes.js'
 
 export class JsonParsingError extends Error {
   constructor(message, codeFrame, stack) {
     super(message)
     this.name = 'JsonParsingError'
+    this.code = ErrorCodes.E_JSON_PARSE
     if (codeFrame) {
       this.codeFrame = codeFrame
     }
